backend/requirements.txt
openai-whisper\ngTTS\nlibrosa\ntorch\ntorchaudio
\nopenai\n
google-api-python-client\ngoogle-auth-httplib2\ngoogle-auth-oauthlib
